Save game system
Save system works by registering persistable objects with
unique names that have a serialize and deserialize method
for storing the game data or restoring it to individual
objects.

Currently only the player data is persisted but the system
can be easily expanded to scene/map data when such state
is available.

Currently the system only supports a single save slot.